Report compaction job completed status update
