+++ java/compaction-job-execution/src/test/java/sleeper/compaction/jobexecution/CompactSortedFilesReportingTest.java
@@ -16,8 +16,11 @@
 package sleeper.compaction.jobexecution;
 
 import org.junit.Test;
+import org.mockito.InOrder;
+import org.mockito.Mockito;
 import sleeper.compaction.job.CompactionJob;
 import sleeper.compaction.job.CompactionJobStatusStore;
+import sleeper.compaction.job.CompactionJobSummary;
 import sleeper.compaction.jobexecution.testutils.CompactSortedFilesTestBase;
 import sleeper.compaction.jobexecution.testutils.CompactSortedFilesTestDataHelper;
 import sleeper.core.record.Record;
@@ -28,8 +31,6 @@ import sleeper.statestore.StateStore;
 import java.util.List;
 
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static sleeper.compaction.jobexecution.testutils.CompactSortedFilesTestData.keyAndTwoValuesSortedEvenLongs;
 import static sleeper.compaction.jobexecution.testutils.CompactSortedFilesTestData.keyAndTwoValuesSortedOddLongs;
 import static sleeper.compaction.jobexecution.testutils.CompactSortedFilesTestUtils.createCompactSortedFiles;
@@ -57,11 +58,14 @@ public class CompactSortedFilesReportingTest extends CompactSortedFilesTestBase 
         dataHelper.addFilesToStateStoreForJob(compactionJob);
 
         // When
-        createCompactSortedFiles(schema, compactionJob, stateStore, jobStatusStore).compact();
+        CompactionJobSummary summary =
+                createCompactSortedFiles(schema, compactionJob, stateStore, jobStatusStore).compact();
 
         // Then
-        verify(jobStatusStore).jobStarted(compactionJob);
-        verifyNoMoreInteractions(jobStatusStore);
+        InOrder order = Mockito.inOrder(jobStatusStore);
+        order.verify(jobStatusStore).jobStarted(compactionJob);
+        order.verify(jobStatusStore).jobCompleted(compactionJob, summary);
+        order.verifyNoMoreInteractions();
     }
 
 }
